refactor(#58) :: mail 폴더 구조 리팩토링
