+++ src/entities/mail/constants/SidebarMenuTitle.ts
