+++ src/entities/mail/types/MailModalProps.ts
