+++ src/features/Home/Sidebar/index.ts
@@ -1,0 +1,4 @@
+export { default as MiniCalendar } from './MiniCalendar';
+export { default as Display } from './Display';
+export { default as ToDoList } from './Todo';
+export { default as Category } from './Category';
