+++ src/features/Mail/index.ts
@@ -1,0 +1,2 @@
+export { default as Sidebar } from './Sidebar';
+export { default as Content } from './Content';
