+++ src/features/Mail/ui/Image.tsx
