+++ src/features/Mail/ui/WriteIcon.tsx
