--- src/pages/Mail/ui/TrashcanIcon.tsx
@@ -1,21 +1,0 @@
-const TrashcanIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="32"
-      height="32"
-      viewBox="0 0 32 32"
-      fill="none"
-    >
-      <path
-        d="M19.6533 12L19.192 24M12.808 24L12.3467 12M25.6373 7.72C26.0933 7.78933 26.5467 7.86267 27 7.94133M25.6373 7.72L24.2133 26.2307C24.1552 26.9843 23.8148 27.6882 23.26 28.2017C22.7053 28.7151 21.9772 29.0002 21.2213 29H10.7787C10.0228 29.0002 9.2947 28.7151 8.73997 28.2017C8.18525 27.6882 7.84479 26.9843 7.78667 26.2307L6.36267 7.72M25.6373 7.72C24.0985 7.48736 22.5517 7.3108 21 7.19067M6.36267 7.72C5.90667 7.788 5.45333 7.86133 5 7.94M6.36267 7.72C7.90151 7.48736 9.44831 7.3108 11 7.19067M21 7.19067V5.96933C21 4.396 19.7867 3.084 18.2133 3.03467C16.7382 2.98752 15.2618 2.98752 13.7867 3.03467C12.2133 3.084 11 4.39733 11 5.96933V7.19067M21 7.19067C17.6716 6.93344 14.3284 6.93344 11 7.19067"
-        stroke="#5A5A5A"
-        stroke-width="2"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-      />
-    </svg>
-  );
-};
-
-export default TrashcanIcon;
